del cache when exit, but not when editting draft

+++ tkpdseller/src/main/java/com/tokopedia/seller/product/edit/view/fragment/ProductAddFragment.java
@@ -587,7 +587,7 @@ public class ProductAddFragment extends BaseDaggerFragment implements ProductAdd
                 ImageSelectModel imageSelectModel = imagesSelectView.getSelectedImage();
                 if (imageSelectModel!= null) {
                     String path = imageSelectModel.getUriOrPath();
-                    if (!TextUtils.isEmpty(path)) {
+                    if (!TextUtils.isEmpty(path) && getProductDraftId() <= 0) {
                         FileUtils.deleteAllCacheTkpdFile(path);
                     }
                 }
@@ -601,7 +601,14 @@ public class ProductAddFragment extends BaseDaggerFragment implements ProductAdd
     public void onImageEditor(String uriOrPath) {
         ArrayList<String> imageUrls = new ArrayList<>();
         imageUrls.add(uriOrPath);
-        ImageEditorActivity.start(getContext(), ProductAddFragment.this, imageUrls, null);
+        ImageEditorActivity.start(getContext(), ProductAddFragment.this, imageUrls, null, getProductDraftId() <= 0 );
+    }
+
+    @Override
+    public void onRemovePreviousPath(String uri) {
+        if (!TextUtils.isEmpty(uri) && getProductDraftId() <= 0) {
+            FileUtils.deleteAllCacheTkpdFile(uri);
+        }
     }
 
     private void clearFocus(){
